Add comment delete component.

+++ src/main/java/io/plyschik/springbootblog/controller/CommentController.java
@@ -1,6 +1,5 @@
 package io.plyschik.springbootblog.controller;
 
-import io.plyschik.springbootblog.dto.Alert;
 import io.plyschik.springbootblog.dto.CommentDto;
 import io.plyschik.springbootblog.dto.PostsCommentApiResponse;
 import io.plyschik.springbootblog.exception.CommentNotFoundException;
@@ -11,8 +10,6 @@ import io.plyschik.springbootblog.service.PostService;
 import io.plyschik.springbootblog.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
-import org.springframework.context.i18n.LocaleContextHolder;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -22,8 +19,6 @@ import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.ModelAndView;
-import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.validation.Valid;
 
@@ -96,63 +91,18 @@ public class CommentController {
     }
 
     @PreAuthorize("hasPermission(#commentId, 'Comment', 'delete')")
-    @GetMapping("/posts/{postId}/comments/{commentId}/delete")
-    public ModelAndView deleteConfirmation(
-        @PathVariable long postId,
-        @PathVariable long commentId,
-        RedirectAttributes redirectAttributes
-    ) {
+    @DeleteMapping("/api/posts/{postId}/comments/{commentId}")
+    public ResponseEntity<Void> delete(@PathVariable Long postId, @PathVariable Long commentId) {
         try {
-            ModelAndView modelAndView = new ModelAndView("comment/delete");
-            modelAndView.addObject("postId", postId);
-            modelAndView.addObject("comment", commentService.getComment(commentId));
+            if (!postService.existsById(postId)) {
+                return ResponseEntity.notFound().build();
+            }
 
-            return modelAndView;
-        } catch (CommentNotFoundException exception) {
-            redirectAttributes.addFlashAttribute(
-                "alert",
-                new Alert("danger", messageSource.getMessage(
-                    "message.comment_not_found",
-                    null,
-                    LocaleContextHolder.getLocale()
-                ))
-            );
-
-            return new ModelAndView(String.format("redirect:/posts/%d", postId));
-        }
-    }
-
-    @PreAuthorize("hasPermission(#commentId, 'Comment', 'delete')")
-    @PostMapping("/posts/{postId}/comments/{commentId}/delete")
-    public ModelAndView delete(
-        @PathVariable long postId,
-        @PathVariable long commentId,
-        RedirectAttributes redirectAttributes
-    ) {
-        try {
             commentService.deleteComment(commentId);
 
-            redirectAttributes.addFlashAttribute(
-                "alert",
-                new Alert("success", messageSource.getMessage(
-                    "message.comment_has_been_successfully_deleted",
-                    null,
-                    LocaleContextHolder.getLocale()
-                ))
-            );
-
-            return new ModelAndView(String.format("redirect:/posts/%d", postId));
-        } catch (EmptyResultDataAccessException exception) {
-            redirectAttributes.addFlashAttribute(
-                "alert",
-                new Alert("danger", messageSource.getMessage(
-                    "message.comment_not_found",
-                    null,
-                    LocaleContextHolder.getLocale()
-                ))
-            );
-
-            return new ModelAndView(String.format("redirect:/posts/%d", postId));
+            return ResponseEntity.noContent().build();
+        } catch (PostNotFoundException | CommentNotFoundException exception) {
+            return ResponseEntity.notFound().build();
         }
     }
 }
